+++ ainer-framework/ainer-test-support/src/test/java/dev/ainer/testsupport/application/TestSupportTestApplication.java
@@ -1,0 +1,12 @@
+package dev.ainer.testsupport.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestSupportTestApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TestSupportTestApplication.class, args);
+    }
+}